Add files via upload
when I attempted to convert with ABBY, the program would not open this file. So I divided it into smaller portions for my code (page range), but the middle range (pages 251-600) in the document will not save. Not worth drilling down too much to find the poisoned root, I think. So here's the converted portions of the report from pages 1-250 and 601-893 in the PDF available from IPCC's archive. feel free to combine into one, and let me know if you can figure out the problem with that middle range!